[#137] refactor: getVoteSummary 메서드 이름을 getVoteResultSummary로 변경
