+++ src/main/java/goorm/eagle7/stelligence/domain/vote/custom/VoteCustomRepository.java
@@ -1,6 +1,7 @@
 package goorm.eagle7.stelligence.domain.vote.custom;
 
-import goorm.eagle7.stelligence.domain.vote.VoteResultSummary;
+import goorm.eagle7.stelligence.domain.vote.model.VoteResultSummary;
+import goorm.eagle7.stelligence.domain.vote.model.VoteSummary;
 
 /**
  * Vote의 데이터를 저장하고 조회하는 Repository 인터페이스 입니다.
@@ -13,5 +14,12 @@ public interface VoteCustomRepository {
 	 * @param contributeId Contribute의 ID
 	 * @return 투표 결과 (총 투표 수, 찬성 투표 수)
 	 */
-	VoteResultSummary getVoteSummary(Long contributeId);
+	VoteResultSummary getVoteResultSummary(Long contributeId);
+
+	/**
+	 * Contribute에 대한 투표 현황을 반환합니다.
+	 * @param contributeId
+	 * @return 투표 현황 (찬성 개수, 반대 개수)
+	 */
+	VoteSummary getVoteSummary(Long contributeId);
 }
