+++ src/test/java/goorm/eagle7/stelligence/domain/contribute/scheduler/ContributeSchedulingActionDeterminerTest.java
@@ -12,8 +12,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import goorm.eagle7.stelligence.domain.contribute.model.Contribute;
 import goorm.eagle7.stelligence.domain.contribute.model.ContributeStatus;
-import goorm.eagle7.stelligence.domain.vote.VoteResultSummary;
 import goorm.eagle7.stelligence.domain.vote.custom.VoteCustomRepository;
+import goorm.eagle7.stelligence.domain.vote.model.VoteResultSummary;
 
 @ExtendWith(MockitoExtension.class)
 class ContributeSchedulingActionDeterminerTest {
@@ -30,7 +30,8 @@ class ContributeSchedulingActionDeterminerTest {
 		Contribute contribute = contribute(1L, null, ContributeStatus.VOTING, null);
 
 		//when
-		when(voteCustomRepository.getVoteSummary(contribute.getId())).thenReturn(new VoteResultSummary(1L, 100, 80));
+		when(voteCustomRepository.getVoteResultSummary(contribute.getId())).thenReturn(
+			new VoteResultSummary(1L, 100, 80));
 		ContributeSchedulingAction action = contributeSchedulingActionDeterminer.check(contribute);
 
 		//then
@@ -43,7 +44,8 @@ class ContributeSchedulingActionDeterminerTest {
 		Contribute contribute = contribute(1L, null, ContributeStatus.VOTING, null);
 
 		//when
-		when(voteCustomRepository.getVoteSummary(contribute.getId())).thenReturn(new VoteResultSummary(1L, 100, 79));
+		when(voteCustomRepository.getVoteResultSummary(contribute.getId())).thenReturn(
+			new VoteResultSummary(1L, 100, 79));
 		ContributeSchedulingAction action = contributeSchedulingActionDeterminer.check(contribute);
 
 		//then
@@ -56,7 +58,8 @@ class ContributeSchedulingActionDeterminerTest {
 		Contribute contribute = contribute(1L, null, ContributeStatus.VOTING, null);
 
 		//when
-		when(voteCustomRepository.getVoteSummary(contribute.getId())).thenReturn(new VoteResultSummary(1L, 100, 29));
+		when(voteCustomRepository.getVoteResultSummary(contribute.getId())).thenReturn(
+			new VoteResultSummary(1L, 100, 29));
 		ContributeSchedulingAction action = contributeSchedulingActionDeterminer.check(contribute);
 
 		//then
